Add proj4js & wicket lib

+++ TerugmelddataWDC.js
@@ -15,7 +15,7 @@
 			dataType: tableau.dataTypeEnum.string
 		}, {
 			id: "meldingsNummer",
-			datatype: tableau.dataTypeEnum.int
+			datatype: tableau.dataTypeEnum.float
 		},
 			{
             id: "location",
